requirements.txt
fastapi
uvicorn
pillow
moondream
requests